First commit with basic functionality
Based on an alist of metadata fields for an org entry,
`org-cc-metadata-format`, which includes instructions on how to
format them, and a custom function to construct these metadata
fields `org-cc-get-metadata`, map across entries using
`org-map-entries`, call `org-cc-metadata-format` at each,
collect the results, create a formated hash table of completion
strings for `completing-read`, call `completing-read` with it,
go to the entry selected, and show the selected entry.
Only on one custom command at a time is supported.